+++ imml/preprocessing/select_incomplete_samples.py
@@ -77,7 +77,6 @@ def select_incomplete_samples(Xs: list):
         mask = torch.stack([torch.isnan(X).any(axis=1) for X in Xs], axis=1)
     else:
         mask = np.stack([pd.isna(X).any(axis=1) for X in Xs], axis=1)
-    mask = np.stack([np.isnan(X).any(axis=1) for X in Xs], axis=1)
     mask = mask.any(axis=1)
     transformed_Xs = [X[mask] for X in Xs]
     return transformed_Xs